# apps/tenant-core/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GSMS Tenant Core API",
    "version": "0.1.0",
    "description": "Identité portail client GSMS : session JWT cookie, organisation, workspaces (établissements), membres. Les routes `/api/bff/*` sont des agrégats mock CRM/Desk en attendant les vrais branchements.\n\nAuth : cookie httpOnly `gsms_session` (JWT HS256)."
  },
  "servers": [
    {
      "url": "http://localhost:3090",
      "description": "Dev local"
    }
  ],
...
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "tags": ["session"],
        "summary": "Connexion",
        "operationId": "login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session créée (Set-Cookie gsms_session)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MeResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/api/auth/me": {
      "get": {
        "tags": ["session"],
        "summary": "Session courante",
        "operationId": "getMe",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": {
            "description": "Profil + org + workspaces + workspace actif",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MeResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "gsms_session"
      }
    },
    "schemas": {
      "ErrorBody": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string" },
          "details": { "type": "object", "additionalProperties": true }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string" },
          "workspaceId": {
            "type": "string",
            "format": "uuid",
            "description": "Optionnel — workspace actif dès le login"
          }
        }
      },
      "User": {
        "type": "object",
        "required": ["id", "email", "name"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": "string", "format": "email" },
          "name": { "type": "string" }
        }
      },
      "Organization": {
        "type": "object",
        "required": ["id", "name"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" }
        }
      },
      "Workspace": {
        "type": "object",
        "required": ["id", "name"],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "workspace_id canonique GSMS"
          },
          "name": { "type": "string" },
          "label": { "type": ["string", "null"] }
        }
      },
      "MeResponse": {
        "type": "object",
        "required": ["user", "organization", "workspaces", "role"],
        "properties": {
          "user": { "$ref": "#/components/schemas/User" },
          "organization": { "$ref": "#/components/schemas/Organization" },
          "workspace": {
            "oneOf": [
              { "$ref": "#/components/schemas/Workspace" },
              { "type": "null" }
            ]
          },
          "workspaces": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Workspace" }
          },
          "role": {
            "type": "string",
            "enum": ["owner", "member", "viewer"]
          }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Cookie session manquant ou invalide",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorBody" }
          }
        }
      },
      "Error": {
        "description": "Erreur métier",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorBody" }
          }
        }
      }
    }
  }
}